chore(tests): cleanup test config, harden native mocks
