--- packages/fiber/__mocks__/expo-file-system.ts
@@ -1,1 +1,0 @@
-export const readAsStringAsync = async () => new Promise((res) => res(''))
