+++ packages/fiber/__mocks__/expo-gl.ts
@@ -1,11 +1,14 @@
 import * as React from 'react'
-import { createWebGLContext } from '@react-three/test-renderer/src/createWebGLContext'
+import type { GLViewProps, ExpoWebGLRenderingContext } from 'expo-gl'
 
-export const GLView = ({ onContextCreate }: { onContextCreate: (gl: any) => void }) => {
+export function GLView({ onContextCreate }: GLViewProps) {
+  const canvas = React.useMemo(() => Object.assign(document.createElement('canvas'), { width: 1280, height: 800 }), [])
+
   React.useLayoutEffect(() => {
-    const gl = createWebGLContext({ width: 1280, height: 800 } as HTMLCanvasElement)
-    onContextCreate(gl)
-  }, [])
+    const gl = canvas.getContext('webgl2') as ExpoWebGLRenderingContext
+    gl.endFrameEXP = () => {}
+    onContextCreate?.(gl)
+  }, [canvas, onContextCreate])
 
   return null
 }
