+++ packages/fiber/tests/canvas.test.tsx
@@ -1,20 +1,16 @@
-// use default export for jest.spyOn
 import React from 'react'
-import { render, RenderResult } from '@testing-library/react'
-
-import { Canvas, act } from '../../src'
+import { render } from '@testing-library/react'
+import { Canvas, act } from '../src'
 
 describe('web Canvas', () => {
   it('should correctly mount', async () => {
-    let renderer: RenderResult = null!
-
-    await act(async () => {
-      renderer = render(
+    const renderer = await act(async () =>
+      render(
         <Canvas>
           <group />
         </Canvas>,
-      )
-    })
+      ),
+    )
 
     expect(renderer.container).toMatchSnapshot()
   })
@@ -22,26 +18,25 @@ describe('web Canvas', () => {
   it('should forward ref', async () => {
     const ref = React.createRef<HTMLCanvasElement>()
 
-    await act(async () => {
+    await act(async () =>
       render(
         <Canvas ref={ref}>
           <group />
         </Canvas>,
-      )
-    })
+      ),
+    )
 
-    expect(ref.current).toBeDefined()
+    expect(ref.current).toBeInstanceOf(HTMLCanvasElement)
   })
 
   it('should correctly unmount', async () => {
-    let renderer: RenderResult = null!
-    await act(async () => {
-      renderer = render(
+    const renderer = await act(async () =>
+      render(
         <Canvas>
           <group />
         </Canvas>,
-      )
-    })
+      ),
+    )
 
     expect(() => renderer.unmount()).not.toThrow()
   })
@@ -49,13 +44,13 @@ describe('web Canvas', () => {
   it('plays nice with react SSR', async () => {
     const useLayoutEffect = jest.spyOn(React, 'useLayoutEffect')
 
-    await act(async () => {
+    await act(async () =>
       render(
         <Canvas>
           <group />
         </Canvas>,
-      )
-    })
+      ),
+    )
 
     expect(useLayoutEffect).not.toHaveBeenCalled()
   })
