+++ packages/shared/setupTests.ts
@@ -1,3 +1,6 @@
+import * as THREE from 'three'
+import { createWebGLContext } from '@react-three/test-renderer/src/createWebGLContext'
+import { extend } from '@react-three/fiber'
 import 'regenerator-runtime/runtime'
 
 declare global {
@@ -35,3 +38,23 @@ if (!global.PointerEvent) {
     getPredictedEvents = () => []
   }
 }
+
+// Polyfills WebGL canvas
+function getContext(contextId: '2d', options?: CanvasRenderingContext2DSettings): CanvasRenderingContext2D | null
+function getContext(
+  contextId: 'bitmaprenderer',
+  options?: ImageBitmapRenderingContextSettings,
+): ImageBitmapRenderingContext | null
+function getContext(contextId: 'webgl', options?: WebGLContextAttributes): WebGLRenderingContext | null
+function getContext(contextId: 'webgl2', options?: WebGLContextAttributes): WebGL2RenderingContext | null
+function getContext(contextId: string): RenderingContext | null {
+  if (contextId === 'webgl' || contextId === 'webgl2') {
+    return createWebGLContext(this)
+  }
+  return null
+}
+
+HTMLCanvasElement.prototype.getContext = getContext
+
+// Extend catalogue for render API in tests
+extend(THREE)
